backend/pyproject.toml
[project]
name = "autogen-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen>=0.9.3",
    "fastapi>=0.115.13",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
    "scholarly>=1.7.11",
    "streamlit>=1.46.0",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
]
